Created SocialUser
Bookmarklet designed to extract user IDs from various social media platforms, including Twitter, Facebook, Instagram, and TikTok. It begins by checking the current URL to determine if it belongs to one of the supported platforms. If the website is supported, it extracts the user ID from the page content and alerts the user with the corresponding platform's User ID. If the website is not supported, it alerts the user that the website is not supported. The script utilises different methods for extracting user IDs based on the structure of each platform's web page.